chore(db): address review — bigint counters, import.meta.url test path
Declare moved/dropped as BIGINT to match count(*); resolve the migration
file relative to import.meta.url so the static test survives a root-level
vitest invocation; renumber the ADR-019 anchors the removed import shifted.

+++ infra/terraform/modules/gke-mcp/lore-platform/charts/ui-helm/migrations/0035_migrate_legacy_org_shared_chunks.sql
@@ -54,8 +54,8 @@
 DO $$
 DECLARE
   r RECORD;
-  moved INTEGER;
-  dropped INTEGER;
+  moved BIGINT;
+  dropped BIGINT;
 BEGIN
   FOR r IN
     SELECT rep.full_name, rep.team
